+++ test/mock/normal/src/http/head-head/index.js
@@ -1,0 +1,11 @@
+exports.handler = async (event) => {
+  const body = event
+  body.message = 'Hello from head /head'
+  return {
+    statusCode: 200,
+    headers: {
+      'content-type': 'application/json',
+      body: JSON.stringify(body)
+    }
+  }
+}