+++ changelog.md
@@ -6,8 +6,10 @@
 
 ### Added
 
-- Add support for `@http` catchall syntax (e.g. `get /api/*`)
-- Adds experimental support for `@proxy`
+- Added support for `@http` catchall syntax (e.g. `get /api/*`)
+- Added support for `@http` `any` method syntax (e.g. `any /path`)
+- Added support for `@http` `head` + `options` methods
+- Added experimental support for `@proxy`
 
 ### Changed